Rename location and node to components
Location and Node were originally created as entities, but this might be
misleading. While they will be instantiated in the game, they have no
behavior themselves. To avoid any confusion, they have been moved into a
new module for components.
