+++ game/simulation/src/component/mod.rs
