+++ game/simulation/src/component/node.rs
