+++ game/simulation/src/lib.rs
@@ -14,6 +14,6 @@
 //! which are passed to the simulation as arguments.
 
 pub mod bus;
-pub mod entity;
+pub mod component;
 
 const TILE_SIZE: u32 = 64;
